widgets are now a feature of the explore page
