+++ KlockWork-iOS/KlockWork-iOS/SharedViews/Widget/Widget.swift
@@ -1,0 +1,12 @@
+//
+//  Widget.swift
+//  KlockWork-iOS
+//
+//  Created by Ryan Priebe on 2024-06-10.
+//
+
+import SwiftUI
+
+struct Widget {
+    
+}
